+++ packages/editor-common/web/src/Utils/overflowUtils.ts
@@ -21,23 +21,22 @@ const getElementCoordsInWindow = (elem: HTMLElement): ElementCoords => {
   return { top: Math.round(top), left: Math.round(left) };
 };
 
-const getWidth = (element: HTMLElement): number => {
-  return element.clientWidth;
-};
-
-export const isElementOutOfWindow = (element: HTMLElement): boolean | undefined => {
-  if (element && document) {
-    const modalOffset = getElementCoordsInWindow(element).left;
-    const modalWidth = getWidth(element);
-    const rootEditor = element.closest('[data-hook=root-editor]') as HTMLElement;
-    const editorOffset = rootEditor && getElementCoordsInWindow(rootEditor).left;
-    const editorWidth = (rootEditor && rootEditor.getBoundingClientRect().width) || 999999;
-    if (modalWidth + (modalOffset - editorOffset) > editorWidth) {
-      return true;
-    } else {
-      return false;
-    }
-  } else {
-    console.error('window or document not found'); //eslint-disable-line
-  }
+export const elementOverflowWithEditor = (element: HTMLElement): Record<string, number> => {
+  const elementCoordsInWindow = getElementCoordsInWindow(element);
+  const elementOffsetLeft = elementCoordsInWindow.left;
+  const elementOffsetTop = elementCoordsInWindow.top;
+  const elementWidth = element.clientWidth;
+  const elementHeight = element.clientHeight;
+  const rootEditor = element.closest('[data-hook=ricos-editor]') as HTMLElement;
+  const rootEditorCoordsInWindow = rootEditor && getElementCoordsInWindow(rootEditor);
+  const rootEditorOffsetLeft = rootEditorCoordsInWindow.left;
+  const rootEditorOffsetTop = rootEditorCoordsInWindow.top;
+  const editorWidth = (rootEditor && rootEditor.getBoundingClientRect().width) || 999999;
+  const editorHeight = (rootEditor && rootEditor.getBoundingClientRect().height) || 999999;
+  const overflowRight = elementWidth + (elementOffsetLeft - rootEditorOffsetLeft) - editorWidth;
+  const overflowBottom = elementHeight + (elementOffsetTop - rootEditorOffsetTop) - editorHeight;
+  return {
+    overflowRight: overflowRight > 0 ? overflowRight : 0,
+    overflowBottom: overflowBottom > 0 ? overflowBottom : 0,
+  };
 };