+++ packages/editor-common/web/src/Utils/tiptap/index.ts
@@ -1,3 +1,0 @@
-export { createNodeExtension } from './createNodeExtension';
-export { createMarkExtension } from './createMarkExtension';
-export { createGenericExtension } from './createGenericExtension';