+++ packages/editor-common/web/src/index.ts
@@ -133,7 +133,7 @@ export {
 import DraftOffsetKey from '@wix/draft-js/lib/DraftOffsetKey';
 export { DraftOffsetKey };
 
-export { isElementOutOfWindow } from './Utils/overflowUtils';
+export { elementOverflowWithEditor } from './Utils/overflowUtils';
 
 export const emptyDraftContent = {
   entityMap: {},
@@ -151,4 +151,3 @@ export const emptyDraftContent = {
 };
 
 export { getSelectionStyles } from './Utils/inlineStyleUtils';
-export { createNodeExtension, createMarkExtension, createGenericExtension } from './Utils/tiptap';