+++ packages/editor/web/src/RichContentEditor/EditorCommands.ts
@@ -1,3 +1,4 @@
+import { pick } from 'lodash';
 import {
   EditorState,
   getColor,
@@ -25,6 +26,9 @@ import {
   removeLinksInSelection,
   triggerMention,
   insertMention,
+  indentSelectedBlocks,
+  mergeBlockData,
+  getAnchorBlockData,
   getAnchorableBlocks,
 } from 'wix-rich-content-editor-common';
 import {
@@ -120,6 +124,8 @@ const TO_DRAFT_PLUGIN_TYPE_MAP = {
   [MAP_TYPE]: MAP_TYPE,
   [ACTION_BUTTON_TYPE]: ACTION_BUTTON_TYPE,
   [LINK_BUTTON_TYPE]: LINK_BUTTON_TYPE,
+  [RICOS_INDENT_TYPE]: RICOS_INDENT_TYPE,
+  [RICOS_LINE_SPACING_TYPE]: RICOS_LINE_SPACING_TYPE,
 };
 
 const TO_RICOS_PLUGIN_TYPE_MAP = {
@@ -166,6 +172,8 @@ const insertDecorationsMap = {
   [RICOS_MENTION_TYPE]: insertMention,
   [RICOS_TEXT_COLOR_TYPE]: setTextColor,
   [RICOS_TEXT_HIGHLIGHT_TYPE]: setHighlightColor,
+  [RICOS_INDENT_TYPE]: indentSelectedBlocks,
+  [RICOS_LINE_SPACING_TYPE]: mergeBlockData,
 };
 
 const deleteDecorationsMapFuncs = {
@@ -173,6 +181,9 @@ const deleteDecorationsMapFuncs = {
   [RICOS_TEXT_COLOR_TYPE]: setTextColor,
   [RICOS_TEXT_HIGHLIGHT_TYPE]: setHighlightColor,
 };
+
+let savedEditorState;
+let savedSelectionState;
 
 export const createEditorCommands = (
   createPluginsDataMap,
@@ -207,6 +218,11 @@ export const createEditorCommands = (
     getSelectedData: EditorCommands['getSelectedData'];
     getSelectedType: EditorCommands['getSelectedType'];
     getPluginsList: EditorCommands['getPluginsList'];
+    getBlockSpacing: EditorCommands['getBlockSpacing'];
+    saveEditorState: EditorCommands['saveEditorState'];
+    loadEditorState: EditorCommands['loadEditorState'];
+    saveSelectionState: EditorCommands['saveSelectionState'];
+    loadSelectionState: EditorCommands['loadSelectionState'];
   } = {
     getSelection: () => {
       const { isCollapsed, getHasFocus } = getEditorState().getSelection();
@@ -226,6 +242,18 @@ export const createEditorCommands = (
     getLinkDataInSelection: () => getLinkDataInSelection(getEditorState()),
     getSelectedData: () => getEntityData(getEditorState()) || {},
     getSelectedType: () => getEntityType(getEditorState()) || {},
+    getBlockSpacing: () => {
+      const { dynamicStyles = {} } = getAnchorBlockData(getEditorState());
+      return pick(dynamicStyles, ['line-height', 'padding-top', 'padding-bottom']);
+    },
+    saveEditorState: () => (savedEditorState = getEditorState()),
+    loadEditorState: () => {
+      const selection = savedEditorState.getSelection();
+      setEditorState(EditorState.forceSelection(savedEditorState, selection));
+    },
+    saveSelectionState: () => (savedSelectionState = getEditorState().getSelection()),
+    loadSelectionState: () =>
+      setEditorState(EditorState.forceSelection(getEditorState(), savedSelectionState)),
     getPluginsList: settings => {
       const { isRicosSchema } = settings || {};
       const pluginsList = plugins?.map(plugin =>