+++ packages/plugin-divider/web/src/tiptap/tiptap.ts
@@ -1,0 +1,14 @@
+import { CreateRicosExtensions } from 'wix-rich-content-common';
+import { Divider as Component } from './component';
+
+export const createTiptapExtensions: CreateRicosExtensions = defaultOptions => [
+  {
+    type: 'node' as const,
+    Component,
+    createComponentDataDefaults: ({ DividerData }) => DividerData.fromJSON({}),
+    createExtensionConfig: () => ({
+      name: 'divider',
+      defaultOptions,
+    }),
+  },
+];