+++ packages/plugin-map/web/src/toolbar/inline-buttons.tsx
@@ -84,6 +84,7 @@ const createInlineButtons: CreateInlineButtons = ({
     {
       keyName: 'settings',
       type: BUTTONS.EXTERNAL_MODAL,
+      fullHeight: true,
       icon: icons.settings || PluginSettingsIcon,
       modalElement: MapSettingsModal,
       modalStyles: getModalStyles({ isMobile }),