+++ packages/ricos-editor/web/src/modals/mergeModalStyles.ts
@@ -9,22 +9,26 @@ const modalStyleDefaults = {
   },
 };
 
-const mergeModalStyles = (modalStyles, theme) => {
+const mergeModalStyles = (modalStyles, theme, fullHeight) => {
   // Render modal below top toolbar of BM sites
   const headerHeight = document.querySelector('[data-hook=bsm-main-container] [data-hook=header]')
     ?.clientHeight;
+  const styles = fullHeight && headerHeight ? { marginTop: headerHeight } : {};
+
   return {
     content: Object.assign(
       {},
       (modalStyles || modalStyleDefaults).content,
-      { zIndex: 20001, ...((headerHeight && { paddingTop: headerHeight }) || {}) },
-      theme?.modalTheme?.content || {}
+      { zIndex: 20001 },
+      theme?.modalTheme?.content || {},
+      styles
     ),
     overlay: Object.assign(
       {},
       (modalStyles || modalStyleDefaults).overlay,
       { zIndex: 666666 },
-      theme?.modalTheme?.overlay || {}
+      theme?.modalTheme?.overlay || {},
+      styles
     ),
   };
 };