+++ CHANGELOG.md
@@ -17,8 +17,104 @@
     Click to see more.
   </summary>
 
+### :bug: Bug Fix
+
+- `plugin-common`
+  - [#2842](https://github.com/wix/ricos/pull/2842) fix link toolbar shortcut
+
 </details>
 <hr/>
+
+## 8.48.4 (Aug 10, 2021)
+
+### :bug: Bug Fix
+
+- `collapsible-list`
+
+  - [#2831](https://github.com/wix/ricos/pull/2831) add collapsible_list to INSERT_PLUGIN_BUTTONS
+
+- `ricos-viewer` `ricos-editor`
+  - [#2840](https://github.com/wix/ricos/pull/2840) expose `iframeSandboxDomain` prop on editor & viewer fix
+  -
+
+## 8.48.3 (Aug 09, 2021)
+
+### :bug: Bug Fix
+
+- `link`
+  - [#2830](https://github.com/wix/ricos/pull/2830) fix modal design when opened using image toolbar
+
+### :nail_care: Polish
+
+- `ricos-viewer` `ricos-editor`
+  - [#2836](https://github.com/wix/ricos/pull/2836) expose `iframeSandboxDomain` prop on editor & viewer
+
+### :house: Internal
+
+- `editor-common`
+  - [#2824](https://github.com/wix/ricos/pull/2824) overflowUtils - support overflow from bottom + change returned value
+
+## 8.48.2 (Aug 05, 2021)
+
+### :rocket: New Feature
+
+- `*`
+  - [#2817](https://github.com/wix/ricos/pull/2817) [performance] add loading=lazy attribute to images and iframes (lazyImagesAndIframes experiment)
+
+### :house: Internal
+
+- `tiptap`
+  - [#2815](https://github.com/wix/ricos/pull/2815) external `addNodeViewHOC` for generic extensions
+  - [#2825](https://github.com/wix/ricos/pull/2825) tiptap focus extensions
+
+### :bug: Bug Fix
+
+- `link`
+  - [#2822](https://github.com/wix/ricos/pull/2822) fix default checkboxes and add placeholder to link input
+- `*`
+  - [#2826](https://github.com/wix/ricos/pull/2826) fix full height modals overrides by bm's header
+- `link-preview`
+  - [#2827](https://github.com/wix/ricos/pull/2827) fix verticals embed svg's width & height
+
+## 8.48.1 (Aug 03, 2021)
+
+### :bug: Bug Fix
+
+- `ricos-common`
+  - [#2831](https://github.com/wix/ricos/pull/2831) add collapsible_list plugin to INSERT_PLUGIN_BUTTONS
+- `editor`
+  - [#2813](https://github.com/wix/ricos/pull/2813) fix rel value not applying in link panel by config
+
+### :house: Internal
+
+- `tiptap`
+  - [#2800](https://github.com/wix/ricos/pull/2800) image plugin endpoint for tiptap editor
+  - [#2814](https://github.com/wix/ricos/pull/2814) `ExtensionBuilder` + divider refactor
+- `image`
+  - [#2800](https://github.com/wix/ricos/pull/2800) image plugin endpoint for tiptap editor
+- `social-polls`
+  - [#2809](https://github.com/wix/ricos/pull/2809) revert loadable support (fixes css issues)
+
+## 8.48.0 (Aug 03, 2021)
+
+### :rocket: New Feature
+
+- `ricos-content`
+  - [#2804](https://github.com/wix/ricos/pull/2804) extractAppEmbedData
+
+### :nail_care: Polish
+
+- `schema`
+  - [#2662](https://github.com/wix/ricos/pull/2662) FileSource contains `privacy`, `id` fields
+  - [#2662](https://github.com/wix/ricos/pull/2662) FileSource's `custom` field deprecation
+
+### :bug: Bug Fix
+
+- `ricos-content`
+  - [#2812](https://github.com/wix/ricos/pull/2812) fix fromHtml/ck-editor/parser span rules
+- `link`
+  - [#2802](https://github.com/wix/ricos/pull/2802) fix modal design when opened using external toolbar
+  - [#2803](https://github.com/wix/ricos/pull/2803) fix default value for newTab checkbox (true by default)
 
 ## 8.47.15 (Aug 01, 2021)